+++ simulator/player.h
@@ -20,14 +20,14 @@ public:
   player();
   card seeCard(card a, Rng& rng);
   void resetCount();
-  int getBet();
+  long getBet();
   action getBasicStrategyAction(gamestate& a);
   action getAction(gamestate& g, int stackIndex); //for a stack in a gamestate, will he hit, stay, surrender, etc.
   bool takesInsurance();
   float getTrueCount();
-  float getBankroll();
-  void getMoney(float a);
-  void loseMoney(float a);
+  long getBankroll();
+  void getMoney(long a);
+  void loseMoney(long a);
   void setCountingMethod(countingMethod a);
   void setConfusionMatrix(c_matrix& a);
   void resetAll();
